+++ src/extractor/scripting_environment_lua.cpp
@@ -870,7 +870,7 @@ void Sol2ScriptingEnvironment::ProcessElements(
     const ExtractionRelationContainer &relations,
     std::vector<std::pair<const osmium::Node &, ExtractionNode>> &resulting_nodes,
     std::vector<std::pair<const osmium::Way &, ExtractionWay>> &resulting_ways,
-    std::vector<InputConditionalTurnRestriction> &resulting_restrictions,
+    std::vector<InputTurnRestriction> &resulting_restrictions,
     std::vector<InputManeuverOverride> &resulting_maneuver_overrides)
 {
     ExtractionNode result_node;